server/requirements.txt
fastapi==0.115.8
uvicorn==0.34.0
pymysql==1.1.1
numpy==2.2.2
xgboost==2.1.4
pytz==2024.2
requests==2.32.3
pywebpush==2.0.3
python-dotenv==1.0.1
